# Improve `BaseBuilder::updateBatch()` SQL