--- a/system/Database/MySQLi/Builder.php
+++ b/system/Database/MySQLi/Builder.php
@@ -53,4 +53,49 @@ class Builder extends BaseBuilder
 
         return implode(', ', $this->QBFrom);
     }
+
+    /**
+     * Generates a platform-specific batch update string from the supplied data
+     */
+    protected function _updateBatch(string $table, array $values, string $index): string
+    {
+        $keys = array_keys(current($values));
+
+        // make array for future use with composite keys - `field`
+        // future: $this->QBOptions['constraints']
+        $constraints = [$index];
+
+        // future: $this->QBOptions['updateFields']
+        $updateFields = array_filter($keys, static fn ($index) => ! in_array($index, $constraints, true));
+
+        $sql = 'UPDATE ' . $this->compileIgnore('update') . $table . " AS t\n";
+
+        $sql .= 'INNER JOIN (' . "\n";
+
+        $sql .= implode(
+            " UNION ALL\n",
+            array_map(
+                static fn ($value) => 'SELECT ' . implode(', ', array_map(
+                    static fn ($key, $index) => $index . ' ' . $key,
+                    $keys,
+                    $value
+                )),
+                $values
+            )
+        ) . "\n";
+
+        $sql .= ') u' . "\n";
+
+        $sql .= 'ON ' . implode(
+            ' AND ',
+            array_map(static fn ($key) => 't.' . $key . ' = u.' . $key, $constraints)
+        ) . "\n";
+
+        $sql .= 'SET' . "\n";
+
+        return $sql .= implode(
+            ",\n",
+            array_map(static fn ($key) => 't.' . $key . ' = u.' . $key, $updateFields)
+        );
+    }
 }