--- a/system/Database/MySQLi/Builder.php
+++ b/system/Database/MySQLi/Builder.php
@@ -12,6 +12,8 @@
 namespace CodeIgniter\Database\MySQLi;
 
 use CodeIgniter\Database\BaseBuilder;
+use CodeIgniter\Database\Exceptions\DatabaseException;
+use CodeIgniter\Database\RawSql;
 
 /**
  * Builder for MySQLi
@@ -60,8 +62,10 @@ class Builder extends BaseBuilder
     protected function _updateBatch(string $table, array $values, string $index): string
     {
         // this is a work around until the rest of the platform is refactored
-        $this->QBOptions['constraints'] = [$index];
-        $keys                           = array_keys(current($values));
+        if ($index !== '') {
+            $this->QBOptions['constraints'] = [$index];
+        }
+        $keys = array_keys(current($values));
 
         $sql = $this->QBOptions['sql'] ?? '';