--- a/system/Database/OCI8/Builder.php
+++ b/system/Database/OCI8/Builder.php
@@ -13,6 +13,7 @@ namespace CodeIgniter\Database\OCI8;
 
 use CodeIgniter\Database\BaseBuilder;
 use CodeIgniter\Database\Exceptions\DatabaseException;
+use CodeIgniter\Database\RawSql;
 
 /**
  * Builder for OCI8
@@ -234,8 +235,10 @@ class Builder extends BaseBuilder
     protected function _updateBatch(string $table, array $values, string $index): string
     {
         // this is a work around until the rest of the platform is refactored
-        $this->QBOptions['constraints'] = [$index];
-        $keys                           = array_keys(current($values));
+        if ($index !== '') {
+            $this->QBOptions['constraints'] = [$index];
+        }
+        $keys = array_keys(current($values));
 
         $sql = $this->QBOptions['sql'] ?? '';