--- a/user_guide_src/source/database/query_builder/092.php
+++ b/user_guide_src/source/database/query_builder/092.php
@@ -2,27 +2,28 @@
 
 $data = [
     [
-        'title' => 'My title',
-        'name'  => 'My Name 2',
-        'date'  => 'My date 2',
+        'title' => 'Title 1',
+        'name'  => 'Name 1',
+        'date'  => 'Date 1',
     ],
     [
-        'title' => 'Another title',
-        'name'  => 'Another Name 2',
-        'date'  => 'Another date 2',
+        'title' => 'Title 2',
+        'name'  => 'Name 2',
+        'date'  => 'Date 2',
     ],
 ];
 
 $builder->updateBatch($data, 'title');
 /*
  * Produces:
- * UPDATE `mytable` SET `name` = CASE
- * WHEN `title` = 'My title' THEN 'My Name 2'
- * WHEN `title` = 'Another title' THEN 'Another Name 2'
- * ELSE `name` END,
- * `date` = CASE
- * WHEN `title` = 'My title' THEN 'My date 2'
- * WHEN `title` = 'Another title' THEN 'Another date 2'
- * ELSE `date` END
- * WHERE `title` IN ('My title','Another title')
+ * UPDATE `mytable`
+ * INNER JOIN (
+ * SELECT 'Title 1' `title`, 'Name 1' `name`, 'Date 1' `date` UNION ALL
+ * SELECT 'Title 2' `title`, 'Name 2' `name`, 'Date 2' `date`
+ * ) u
+ * ON `mytable`.`title` = u.`title`
+ * SET
+ * `mytable`.`title` = u.`title`,
+ * `mytable`.`name` = u.`name`,
+ * `mytable`.`date` = u.`date`
  */