@kenjis 2022-08-13
> What version are we running?

Just SQLite3.

> SQLite3 via the SQLite3 driver
https://codeigniter4.github.io/CodeIgniter4/intro/requirements.html