@kenjis 2022-08-13
`$this->db->getVersion()` returns versionString like `3.5.9`. It is not a valid number string.
https://www.php.net/manual/en/sqlite3.version.php
I don't know `(float)` works always as you expect.