@sclubricants 2022-08-13
I tested casting to float. It basically seems to knock everything off from the second decimal. 

`$this->db->getVersion() ` could return an int.. would have been more useful.

https://www.php.net/manual/en/sqlite3.version.php