@kenjis 2022-08-13
> I tested casting to float. It basically seems to knock everything off from the second decimal.

It seems the behavior is not documented.

Unless it is documented, we should not rely on the current behavior.
It is better to use `version_compare()`.
https://www.php.net/manual/en/function.version-compare.php

Or you can use `SQLite3::version()` directly.