@sclubricants 2022-08-13
Thanks, didn't know about that function.