@sclubricants 2022-08-26
https://github.com/codeigniter4/CodeIgniter4/pull/6407 has a slightly better implementation as most of the query is cached and doesn't need rebuilt for each batch run.

Also, while performance is important, there are other considerations in the design of the new query. It allows multiple keys to update on where the current implementation only allows one. The majority of the tables I work with have composite keys so this is very helpful to me. Another thing is that you can use a query as a datasource. Although its not implemented here you could also join a table to the dataset.

I have made quite a few changes in my all inclusive PR https://github.com/codeigniter4/CodeIgniter4/pull/6407

I could update this PR with some of them.