@sclubricants 2022-09-06
> @sclubricants Can you fix PHPCPD error?

Its not as bad as it looks. A single statement can take 11 lines.

```php
            $data = implode(
                " UNION ALL\n",
                array_map(
                    static fn ($value) => 'SELECT ' . implode(', ', array_map(
                        static fn ($key, $index) => $index . ' ' . $key,
                        $keys,
                        $value
                    )),
                    $values
                )
            ) . "\n";
```
The top 6 lines will go away once things get refactored. I wasn't getting this error until I added these.

```php
        // this is a work around until the rest of the platform is refactored
        if ($index !== '') {
            $this->QBOptions['constraints'] = [$index];
        }
        $keys = array_keys(current($values));

```

Can we suppress these somehow?